hos03/requirements.txt
google-adk[database]==1.6.1
litellm==1.74.3